% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetrics.R
\name{EyeMetrics}
\alias{EyeMetrics}
\title{EyeMetrics function
Written by Libby Jenke}
\usage{
EyeMetrics(fixation_data, trials)
}
\arguments{
\item{fixation_data}{A data frame containing the eye tracking data. In the data, each row should correspond to a unique fixation.
The data must contain the following variables: subjects' identification numbers (\code{subject_id}), the trial number of each fixation (\code{trial_number}), AOI numbers of each fixation (\code{aoi_num}; this should be a column that indicates which AOI each fixation was in), and the duration of each fixation (\code{fix_duration}).}

\item{trials}{Indicates the trial(s) that you want output for.}
}
\value{
This function returns a tibble with the following columns: subject ID (\code{subject_id}), AOI number, number of fixations per AOI per subject (\code{num_fix}), total fixation time per AOI per subject (\code{tot_fix_time}), number of returns to each AOI per sujbect (\code{num_returns}), grand total of fixation time across all AOIS for each subject (\code{tot_fix_time_subject}), percentage of time spent on each AOI per subject (\code{percentage_aoi_fix_time}), and the first fixated option per subject (\code{first_aoi}).
}
\description{
This function eye tracking data and calculates useful eye tracking metrics for data analysis.
The metrics include: number of fixations, number of returns, total fixation time (ms), and total fixation percentage for each area of interest (AOI) for each respondent and for each trial in one's eye tracking data.
It also yields the first fixated option for the respondent on each trial.
}
